[DTensor] Fix slow sharding prop for stack
As identified in the original issue, there is quadratic complexity in
the number of input tensors, due to an improperly written sharding prop
rule.

The previous code generated N output strategies for the stack op, one
based on each of the original N input strategies.  However, Each of the
N output strategies was the same.  The heuristic in the stack rule is to
find one of the N inputs and follow that one.

We now just generate one output strategy.

Fixes #169445

ghstack-source-id: 5b7b303289c2f47896e86ae3d08525459486ee7a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/169519